requirements.txt
fastapi
uvicorn
pydantic
httpx
polyline
python-multipart
google-cloud-storage
